LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-at-user-scroll-end-expected.txt


FAIL Tests that window should snap at user scroll end. assert_approx_equals: window.scrollX should be at snapped position. expected 408 +/- 0.5 but got 508
